Remove more redundant code, add KoskiUtil for util functions
